+++ example/shared/src/screens/MapScreen.tsx
@@ -167,7 +167,7 @@ const MapScreenInner = ({
       <AnimatedButton
         activeOpacity={0.6}
         style={[styles.floatingControl, floatingControlStyles]}
-        onPress={() => sheetRef.current?.present(1)}
+        onPress={() => sheetRef.current?.present(0)}
       />
       <ReanimatedTrueSheet
         name="main"